shortest path inverse cost function throughput actual calculation

+++ src/main/java/edu/ucla/cloud/switches/AggregrateSwitch.java
@@ -37,7 +37,7 @@ public class AggregrateSwitch extends Switch {
 	}
 
 	@Override
-	protected int linkCapacity() {
+	protected double linkCapacity() {
 
 		return SwitchConstants.AGGREGATE_THROUGHPUT;
 	}
